+++ HinghwaDict/word/views.py
@@ -86,61 +86,70 @@ def searchWords(request):
 @csrf_exempt
 def manageWord(request, id):
     try:
-        word = Word.objects.get(id=id)
-        if request.method == 'GET':
-            related_words = [word.id for word in word.related_words.all()]
-            related_articles = [article.id for article in word.related_articles.all()]
-            word.views = word.views + 1
-            word.save()
-            user = word.contributor
-            return JsonResponse({"word": {"id": word.id, 'word': word.word, 'definition': word.definition,
-                                          "contributor": {"id": user.id, 'username': user.username,
-                                                          'nickname': user.user_info.nickname,
-                                                          'email': user.email, 'telephone': user.user_info.telephone,
-                                                          'registration_time': user.date_joined.__format__('%Y-%m-%d %H:%M:%S'),
-                                                          'login_time': user.last_login.__format__('%Y-%m-%d %H:%M:%S'),
-                                                          'birthday': user.user_info.birthday,
-                                                          'avatar': user.user_info.avatar,
-                                                          'county': user.user_info.county, 'town': user.user_info.town,
-                                                          'is_admin': user.is_superuser}, "annotation": word.annotation,
-                                          "mandarin": eval(word.mandarin),
-                                          "related_words": related_words, "related_articles": related_articles,
-                                          "views": word.views}}, status=200)
-        elif request.method == 'PUT':
-            body = demjson.decode(request.body)
-            token = request.headers['token']
-            if token_check(token, 'dxw', word.contributor.id):
-                body = body['word']
-                word_form = WordForm(body)
-                for key in body:
-                    if (key in word_form) and len(word_form[key].errors.data):
-                        return JsonResponse({}, status=400)
-                for key in body:
-                    if key != 'related_words' and key != 'related_articles':
-                        setattr(word, key, body[key])
-                    elif key == 'related_words':
-                        word.related_words.clear()
-                        for id in body['related_words']:
-                            wordx = Word.objects.get(id=id)
-                            word.related_words.add(wordx)
-                    elif key == 'related_articles':
-                        word.related_articles.clear()
-                        for id in body['related_articles']:
-                            article = Article.objects.get(id=id)
-                            word.related_articles.add(article)
+        word = Word.objects.filter(id=id)
+        if word.exists():
+            word = word[0]
+            if request.method == 'GET':
+                related_words = [word.id for word in word.related_words.all()]
+                related_articles = [article.id for article in word.related_articles.all()]
+                word.views = word.views + 1
                 word.save()
-                return JsonResponse({}, status=200)
+                user = word.contributor
+                return JsonResponse({"word": {"id": word.id, 'word': word.word, 'definition': word.definition,
+                                              "contributor": {"id": user.id, 'username': user.username,
+                                                              'nickname': user.user_info.nickname,
+                                                              'email': user.email,
+                                                              'telephone': user.user_info.telephone,
+                                                              'registration_time': user.date_joined.__format__(
+                                                                  '%Y-%m-%d %H:%M:%S'),
+                                                              'login_time': user.last_login.__format__(
+                                                                  '%Y-%m-%d %H:%M:%S'),
+                                                              'birthday': user.user_info.birthday,
+                                                              'avatar': user.user_info.avatar,
+                                                              'county': user.user_info.county,
+                                                              'town': user.user_info.town,
+                                                              'is_admin': user.is_superuser},
+                                              "annotation": word.annotation,
+                                              "mandarin": eval(word.mandarin),
+                                              "related_words": related_words, "related_articles": related_articles,
+                                              "views": word.views}}, status=200)
+            elif request.method == 'PUT':
+                body = demjson.decode(request.body)
+                token = request.headers['token']
+                if token_check(token, 'dxw', word.contributor.id):
+                    body = body['word']
+                    word_form = WordForm(body)
+                    for key in body:
+                        if (key in word_form) and len(word_form[key].errors.data):
+                            return JsonResponse({}, status=400)
+                    for key in body:
+                        if key != 'related_words' and key != 'related_articles':
+                            setattr(word, key, body[key])
+                        elif key == 'related_words':
+                            word.related_words.clear()
+                            for id in body['related_words']:
+                                wordx = Word.objects.get(id=id)
+                                word.related_words.add(wordx)
+                        elif key == 'related_articles':
+                            word.related_articles.clear()
+                            for id in body['related_articles']:
+                                article = Article.objects.get(id=id)
+                                word.related_articles.add(article)
+                    word.save()
+                    return JsonResponse({}, status=200)
+                else:
+                    return JsonResponse({}, status=401)
+            elif request.method == 'DELETE':
+                token = request.headers['token']
+                if token_check(token, 'dxw', word.contributor.id):
+                    word.delete()
+                    return JsonResponse({}, status=200)
+                else:
+                    return JsonResponse({}, status=401)
             else:
-                return JsonResponse({}, status=401)
-        elif request.method == 'DELETE':
-            token = request.headers['token']
-            if token_check(token, 'dxw', word.contributor.id):
-                word.delete()
-                return JsonResponse({}, status=200)
-            else:
-                return JsonResponse({}, status=401)
+                return JsonResponse({}, status=405)
         else:
-            return JsonResponse({}, status=405)
+            return JsonResponse({}, status=404)
     except Exception as e:
         return JsonResponse({"msg": str(e)}, status=500)
 
@@ -192,37 +201,42 @@ def searchCharacters(request):
 @csrf_exempt
 def manageCharacter(request, id):
     try:
-        character = Character.objects.get(id=id)
-        if request.method == 'GET':
-            return JsonResponse({'character': {"id": character.id, 'shengmu': character.shengmu, 'ipa': character.ipa,
-                                               'pinyin': character.pinyin, 'yunmu': character.yunmu,
-                                               'shengdiao': character.shengdiao,
-                                               'character': character.character, 'county': character.county,
-                                               'town': character.town}}, status=200)
-        elif request.method == 'PUT':
-            body = demjson.decode(request.body)
-            token = request.headers['token']
-            if token_check(token, 'dxw', -1):
-                body = body['character']
-                character_form = CharacterForm(body)
-                for key in body:
-                    if len(character_form[key].errors.data):
-                        return JsonResponse({}, status=400)
-                for key in body:
-                    setattr(character, key, body[key])
-                character.save()
-                return JsonResponse({}, status=200)
+        character = Character.objects.filter(id=id)
+        if character.exists():
+            character = character[0]
+            if request.method == 'GET':
+                return JsonResponse(
+                    {'character': {"id": character.id, 'shengmu': character.shengmu, 'ipa': character.ipa,
+                                   'pinyin': character.pinyin, 'yunmu': character.yunmu,
+                                   'shengdiao': character.shengdiao,
+                                   'character': character.character, 'county': character.county,
+                                   'town': character.town}}, status=200)
+            elif request.method == 'PUT':
+                body = demjson.decode(request.body)
+                token = request.headers['token']
+                if token_check(token, 'dxw', -1):
+                    body = body['character']
+                    character_form = CharacterForm(body)
+                    for key in body:
+                        if len(character_form[key].errors.data):
+                            return JsonResponse({}, status=400)
+                    for key in body:
+                        setattr(character, key, body[key])
+                    character.save()
+                    return JsonResponse({}, status=200)
+                else:
+                    return JsonResponse({}, status=401)
+            elif request.method == 'DELETE':
+                token = request.headers['token']
+                if token_check(token, 'dxw', -1):
+                    character.delete()
+                    return JsonResponse({}, status=200)
+                else:
+                    return JsonResponse({}, status=401)
             else:
-                return JsonResponse({}, status=401)
-        elif request.method == 'DELETE':
-            token = request.headers['token']
-            if token_check(token, 'dxw', -1):
-                character.delete()
-                return JsonResponse({}, status=200)
-            else:
-                return JsonResponse({}, status=401)
+                return JsonResponse({}, status=405)
         else:
-            return JsonResponse({}, status=405)
+            return JsonResponse({}, status=404)
     except Exception as e:
         return JsonResponse({"msg": str(e)}, status=500)
 
@@ -277,52 +291,57 @@ def searchPronunciations(request):
 @csrf_exempt
 def managePronunciation(request, id):
     try:
-        pronunciation = Pronunciation.objects.get(id=id)
-        if request.method == 'GET':
-            pronunciation.views += 1
-            pronunciation.save()
-            user = pronunciation.contributor
-            return JsonResponse(
-                {"pronunciation": {"id": pronunciation.id, 'word': pronunciation.word.id,
-                                   'source': pronunciation.source,
-                                   'ipa': pronunciation.ipa, 'pinyin': pronunciation.pinyin,
-                                   'contributor': {"id": user.id, 'username': user.username,
-                                                   'nickname': user.user_info.nickname,
-                                                   'email': user.email, 'telephone': user.user_info.telephone,
-                                                   'registration_time': user.date_joined.__format__(
-                                                       '%Y-%m-%d %H:%M:%S'),
-                                                   'login_time': user.last_login.__format__('%Y-%m-%d %H:%M:%S'),
-                                                   'birthday': user.user_info.birthday, 'avatar': user.user_info.avatar,
-                                                   'county': user.user_info.county, 'town': user.user_info.town,
-                                                   'is_admin': user.is_superuser}, 'county': pronunciation.county,
-                                   'town': pronunciation.town, 'visibility': pronunciation.visibility}}, status=200)
-        elif request.method == 'PUT':
-            token = request.headers['token']
-            if token_check(token, 'dxw', pronunciation.contributor.id):
-                body = demjson.decode(request.body)
-                body = body['pronunciation']
-                pronunciation_form = PronunciationForm(body)
-                for key in body:
-                    if (key!='word') and len(pronunciation_form[key].errors.data):
-                        return JsonResponse({}, status=400)
-                for key in body:
-                    if key != 'word':
-                        setattr(pronunciation, key, body[key])
-                    else:
-                        pronunciation.word=Word.objects.get(id=body[key])
+        pronunciation = Pronunciation.objects.filter(id=id)
+        if pronunciation.exists():
+            pronunciation = pronunciation[0]
+            if request.method == 'GET':
+                pronunciation.views += 1
                 pronunciation.save()
-                return JsonResponse({}, status=200)
+                user = pronunciation.contributor
+                return JsonResponse(
+                    {"pronunciation": {"id": pronunciation.id, 'word': pronunciation.word.id,
+                                       'source': pronunciation.source,
+                                       'ipa': pronunciation.ipa, 'pinyin': pronunciation.pinyin,
+                                       'contributor': {"id": user.id, 'username': user.username,
+                                                       'nickname': user.user_info.nickname,
+                                                       'email': user.email, 'telephone': user.user_info.telephone,
+                                                       'registration_time': user.date_joined.__format__(
+                                                           '%Y-%m-%d %H:%M:%S'),
+                                                       'login_time': user.last_login.__format__('%Y-%m-%d %H:%M:%S'),
+                                                       'birthday': user.user_info.birthday,
+                                                       'avatar': user.user_info.avatar,
+                                                       'county': user.user_info.county, 'town': user.user_info.town,
+                                                       'is_admin': user.is_superuser}, 'county': pronunciation.county,
+                                       'town': pronunciation.town, 'visibility': pronunciation.visibility}}, status=200)
+            elif request.method == 'PUT':
+                token = request.headers['token']
+                if token_check(token, 'dxw', pronunciation.contributor.id):
+                    body = demjson.decode(request.body)
+                    body = body['pronunciation']
+                    pronunciation_form = PronunciationForm(body)
+                    for key in body:
+                        if (key != 'word') and len(pronunciation_form[key].errors.data):
+                            return JsonResponse({}, status=400)
+                    for key in body:
+                        if key != 'word':
+                            setattr(pronunciation, key, body[key])
+                        else:
+                            pronunciation.word = Word.objects.get(id=body[key])
+                    pronunciation.save()
+                    return JsonResponse({}, status=200)
+                else:
+                    return JsonResponse({}, status=401)
+            elif request.method == 'DELETE':
+                token = request.headers['token']
+                if token_check(token, 'dxw'):
+                    pronunciation.delete()
+                    return JsonResponse({}, status=200)
+                else:
+                    return JsonResponse({}, status=401)
             else:
-                return JsonResponse({}, status=401)
-        elif request.method == 'DELETE':
-            token = request.headers['token']
-            if token_check(token, 'dxw'):
-                pronunciation.delete()
-                return JsonResponse({}, status=200)
-            else:
-                return JsonResponse({}, status=401)
+                return JsonResponse({}, status=405)
         else:
-            return JsonResponse({}, status=405)
+            return JsonResponse({}, status=404)
     except Exception as e:
         return JsonResponse({"msg": str(e)}, status=500)
